+++ apps/storybook/package.json
@@ -2,6 +2,7 @@
     "name": "storybook",
     "private": true,
     "version": "0.1.0",
+    "type": "module",
     "scripts": {
         "dev": "storybook dev -p 6006",
         "build-storybook": "storybook build",
