+++ apps/storybook/tsconfig.json
@@ -2,9 +2,14 @@
     "extends": "../../packages/typescript-config/react-app.json",
     "compilerOptions": {
         "baseUrl": ".",
+        "module": "ESNext",
+        "moduleResolution": "bundler",
+        "target": "ES2022",
+        "allowSyntheticDefaultImports": true,
+        "esModuleInterop": true,
         "paths": {
-            "blend-v1": ["../../packages/blend/lib/main.ts"],
-            "blend-v1/*": ["../../packages/blend/lib/*"]
+            "@juspay/blend-design-system": ["../../packages/blend/lib/main.ts"],
+            "@juspay/blend-design-system/*": ["../../packages/blend/lib/*"]
         }
     },
     "include": ["stories/**/*", ".storybook/**/*"],
