+++ packages/blend/package.json
@@ -22,7 +22,12 @@
     "scripts": {
         "lint": "eslint .",
         "build": "pnpm lint && tsc --p ./tsconfig.json && vite build",
-        "prepublishOnly": "pnpm run build"
+        "prepublishOnly": "pnpm run build",
+        "test": "vitest",
+        "test:ui": "vitest --ui",
+        "test:coverage": "vitest --coverage",
+        "test:watch": "vitest --watch",
+        "test:run": "vitest run"
     },
     "peerDependencies": {
         "react": "^18.0.0 || ^19.0.0",
@@ -53,24 +58,37 @@
         "@storybook/react": "^8.6.12",
         "@storybook/react-vite": "^8.6.12",
         "@storybook/test": "^8.6.12",
+        "@testing-library/dom": "^10.4.0",
+        "@testing-library/jest-dom": "^6.6.3",
+        "@testing-library/react": "^16.3.0",
+        "@testing-library/react-hooks": "^8.0.1",
+        "@testing-library/user-event": "^14.6.1",
+        "@types/jest-axe": "^3.5.9",
         "@types/node": "^22.15.21",
         "@types/react": "^19.1.2",
         "@types/react-dom": "^19.1.2",
         "@vitejs/plugin-react": "^4.4.1",
+        "@vitest/coverage-v8": "^3.2.4",
+        "@vitest/ui": "^3.2.4",
         "eslint": "^9.25.0",
         "eslint-plugin-react-hooks": "^5.2.0",
         "eslint-plugin-react-refresh": "^0.4.19",
         "eslint-plugin-storybook": "^0.12.0",
         "globals": "^16.0.0",
+        "jest-axe": "^10.0.0",
+        "jest-styled-components": "^7.2.0",
+        "jsdom": "^26.1.0",
         "prettier": "^3.1.0",
         "react": "^19.1.0",
         "react-dom": "^19.1.0",
+        "resize-observer-polyfill": "^1.5.1",
         "storybook": "^8.6.12",
         "typescript": "~5.8.3",
         "typescript-eslint": "^8.30.1",
         "vite": "^6.3.5",
         "vite-plugin-dts": "^4.5.4",
-        "vite-plugin-lib-inject-css": "^2.2.2"
+        "vite-plugin-lib-inject-css": "^2.2.2",
+        "vitest": "^3.2.4"
     },
     "keywords": [
         "react",
